More again7 logging to be able to debug message

+++ testInstructionExecutionEngine/testInstructionExecutionEngine_CloudDB_sendNewTestInstructionsThatIsWaitingToBeSent.go
@@ -752,7 +752,7 @@ func (executionEngine *TestInstructionExecutionEngineStruct) sendTestInstruction
 					ProcessTestInstructionExecutionPubSubRequest_TestInstructionAttributeMessage{
 					TestInstructionAttributeType: fenixExecutionWorkerGrpcApi.
 						ProcessTestInstructionExecutionPubSubRequest_TestInstructionAttributeTypeEnum(
-							testInstructionAttributesForReversed.TestInstructionAttributeType),
+							int32(testInstructionAttributesForReversed.TestInstructionAttributeType)),
 					TestInstructionAttributeUuid:     testInstructionAttributesForReversed.TestInstructionAttributeUuid,
 					TestInstructionAttributeName:     testInstructionAttributesForReversed.TestInstructionAttributeName,
 					AttributeValueAsString:           testInstructionAttributesForReversed.AttributeValueAsString,
